requirements.txt
aniso8601==6.0.0
Click==8.2.1
Flask==1.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.10
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==2.0.3
gunicorn==21.2.0
python-dotenv==1.1.1
